+++ Gemfile
@@ -1,10 +1,4 @@
 source 'https://rubygems.org'
 
-gem "rubyzip"
-gem "opal"
-gem "slop", '~> 4'
-
-gem "rake", group: :test, require: nil
-gem "rspec", group: :test, require: nil
-gem "poltergeist", group: :test, require: nil
-gem "codeclimate-test-reporter", group: :test, require: nil
+gemspec name: 'gamefic'
+gemspec name: 'gamefic-sdk'
