+++ gamefic
@@ -1,4 +1,4 @@
 #!/usr/bin/env ruby
 
-$LOAD_PATH.unshift File.dirname(__FILE__) + "/lib"
-load File.dirname(__FILE__) +  '/bin/gamefic'
+$LOAD_PATH.unshift File.dirname(__FILE__) + '/lib'
+load File.dirname(__FILE__) + '/bin/gamefic'
