+++ gfk
@@ -1,4 +1,4 @@
 #!/usr/bin/env ruby
 
-$LOAD_PATH.unshift File.dirname(__FILE__) + "/lib"
+$LOAD_PATH.unshift File.dirname(__FILE__) + '/lib'
 load File.dirname(__FILE__) + '/bin/gfk'
