--- lib/gamefic/shell/command.rb
@@ -1,4 +1,0 @@
-module Gamefic::Shell::Command
-  autoload :Base, 'gamefic/shell/command/base'
-  autoload :Play, 'gamefic/shell/command/play'
-end
